Option #2 for progress bar update
This option adds the possibility for the user to specify its own handler for the progress update event. If none was specified, we trigger the custom plugin event, else we call the specified function, still passing the current progress percentage as a parameter. 

Usage example for initialization: 
`$(".msf:first").multiStepForm(
            {
                progressHandler: function (progressValue) {
                    $userRegistrationProgressBar.attr('aria-valuenow', progressValue + '%').css('width', progressValue + '%');
                }
            }
        );`